rec Library, extend recAssociate class.
Add recAssociateVec typedef. Use it as return for new GetList member function. This returns a list of all Associated databases.

Add Create member function which creates a new record for a given path.

Use C++11 feature to define table name as a static constexpr variable and use this to replace the pure virtual GetTableName member function.