TFP program, command line file name argument changed.
Since there are at least 2 database types, opening a database needs the window frame to be set appropriately. Instead of just opening a command line argument file name, the name is passed to the Frame constructor.

The TfpFrame::OpenFile function has been split with the function just requesting the filename and passing it to a new TfpFrame::OpenFilename function.